requirements.txt
# ======================================
# BIZNESPLAN GENERATOR - DEPENDENCIES
# ======================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0  # Async PostgreSQL driver (Python 3.13 compatible)
psycopg2-binary==2.9.10  # Sync PostgreSQL driver (for compatibility)
alembic==1.14.0

# AI/LLM
anthropic==0.25.0
langchain==0.1.20
langchain-anthropic==0.1.10
langgraph==0.0.40
langchain-community==0.0.38

# Task Queue
celery==5.3.4
redis==5.0.1

# API Integrations
requests==2.31.0
httpx==0.26.0

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# Utilities
python-dotenv==1.0.0
markdown==3.5.1
validators==0.22.0
python-multipart==0.0.6
jinja2==3.1.2

# Monitoring & Logging
structlog==24.1.0

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3

